Added functionality to job postings save/unsave button
When clicking the save/unsave button, the front end now talks to the database and stores/removes the saved posting information under the selected user.

Note: When the postings/saved tabs are clicked on the job postings page, the saved postings are not updated unless the whole page is refreshed first. Nikki and I are actively trying to fix this